requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
google-api-core==2.24.2
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
uritemplate==4.1.1
urllib3==2.4.0
